renames problem files
