+++ src/components/basicPage/durationtimebar.jsx
